# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WayFarer is a public bus transportation booking server.",
        "version": "1.0.0",
        "title": "Wayfarer APi",
        "termsOfService": "http://swagger.io/terms/",
        
        "license": {
            "name": "MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/buses": {
            "get": {
                "tags": [
                    "Buses"
                ],
                "summary": "Get all buses in the system",
                "description": "An admin can get all the buses",
                "operationId": "getbuses",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Invalid Token"
                    },
                    "200": {
                        "description": "Successfully fetched all bank accounts"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Buses"
                ],
                "summary": "Create a bus",
                "description": "Adds a bus to the system",
                "operationId": "createBus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Bus object needed to create bus",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Bus"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Uauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Bus with this number plate is already registered"
                    },
                    "201": {
                        "description": "Successfully Created an Account"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/trips": {
            "get": {
                "tags": ["Trips"],
                "summary": "Gets all trips in the system",
                "description": "",
                "operationId": "getAllTrips",
                "consumes" : ["application/json"],
                "produces": ["application/json"],
                "parameters":{ 
                    "name": ["origin", "destination"],
                    "in": "path",
                    "description": "Gets all trips, filtered by the search query",
                    "type": "string",
                    "required": false
                },
                "responses": {
                    "400": {
                        "description": "Invalid query strings"
                    },
                    "401" : {
                        "description": "Invalid Token"
                    },
                    "403": {
                        "description": "No Token Provided"
                    },
                    "200": {
                        "description": "All trips retrieved successfully"
                    }

                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": ["Trips"],
                "summary": "Create new trip",
                "description": "",
                "operationId": "createTrip",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": {
                    "in": "body",
                    "name": "newTrip",
                    "description": "Trip Object needed to create a trip",
                    "required": false,
                    "type": "object"
                },
                "responses": {
                    "400":{
                        "description": "Input Validation Error | No token supplied"
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "200": {
                        "description": "Trip created Successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Gets all bookings in the system",
                "description": "",
                "operationId": "getAllBookings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                
                "responses": {
                    "401": {
                        "description": "Invalid Token"
                    },
                    "403": {
                        "description": "User can not see another user's bookings"
                    },
                    "200": {
                        "description": "All bookings retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create new booking",
                "description": "",
                "operationId": "createBooking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": {
                    "in": "body",
                    "name": "newBooking",
                    "description": "Booking Object needed to create a trip",
                    "required": false,
                    "type": "object",
                    "schema": {
                        "$ref":"#/definitions/Booking"
                    }
                },
                "responses": {
                    "400": {
                        "description": "Input Validation Error"
                    },
                    "401": {
                        "description": "No Token provided | Invalid Token"
                    },
                    "422": {
                        "description": "Trip cancelled"
                    },
                    "200": {
                        "description": "Trip created Successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    } 
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "scheme": "bearer",
            "description": "JWT token is received after users register / login,",
            "name": "token",
            "in": "header",
            "bearerFormat": "JWT"
        }
    },
    "definitions": {
        "Bus": {
            "type": "object",
            "required": ["manufacturer","model", "capacity", "year", "number_plate"],
            "properties": {
                "manufacturer": {
                    "type": "string",
                    "example": "Toyota"
                },
                "model": {
                    "type": "string",
                    "example": "Hiace"
                },
                "year": {
                    "type": "number",
                    "example": "2016"
                },
                "capacity": {
                    "type": "number",
                    "example": "22"
                },
                "number_plate": {
                    "type": "string",
                    "example": "IKJ-655TY"
                }
            }
        },
...
        "Booking": {
            "type": "object",
            "required": [
                "trip_id"
            ],
            "properties": {
                "trip_id": {
                    "type": "number",
                    "example": "3"
                },
                "seat_number": {
                    "type": "number",
                    "example": "11"
                }
            }
        }
    },
...
}